feat: remove parquet index
