+++ src/config/src/meta/inverted_index.rs
@@ -15,52 +15,6 @@
 
 use proto::cluster_rpc;
 
-/// Supported inverted index formats:
-///  - Parquet (v2): Index is stored in parquet format
-///  - Tantivy (v3): Index is stored in custom puffin format files
-///  - both: Use both Parquet and Tantivy. Note that this will generate two inverted index files.
-#[derive(Clone, Copy, Debug, Eq, PartialEq, Default)]
-pub enum InvertedIndexFormat {
-    #[default]
-    Parquet,
-    Both,
-    Tantivy,
-}
-
-impl From<&String> for InvertedIndexFormat {
-    fn from(s: &String) -> Self {
-        match s.to_lowercase().as_str() {
-            "both" => InvertedIndexFormat::Both,
-            "tantivy" => InvertedIndexFormat::Tantivy,
-            _ => InvertedIndexFormat::Parquet,
-        }
-    }
-}
-
-impl std::fmt::Display for InvertedIndexFormat {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            InvertedIndexFormat::Parquet => write!(f, "parquet"),
-            InvertedIndexFormat::Both => write!(f, "both"),
-            InvertedIndexFormat::Tantivy => write!(f, "tantivy"),
-        }
-    }
-}
-
-pub enum InvertedIndexTantivyMode {
-    Puffin,
-    Mmap,
-}
-
-impl std::fmt::Display for InvertedIndexTantivyMode {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            InvertedIndexTantivyMode::Puffin => write!(f, "puffin"),
-            InvertedIndexTantivyMode::Mmap => write!(f, "mmap"),
-        }
-    }
-}
-
 #[derive(Clone, Debug, Eq, PartialEq)]
 pub enum InvertedIndexOptimizeMode {
     SimpleSelect(usize, bool),
