+++ src/service/mod.rs
@@ -45,6 +45,7 @@ pub mod promql;
 pub mod ratelimit;
 pub mod schema;
 pub mod search;
+pub mod tantivy;
 pub mod websocket_events;
 
 #[cfg(feature = "enterprise")]
